+++ openair1/PHY/defs_nr_common.h
@@ -230,9 +230,9 @@ typedef struct NR_DL_FRAME_PARMS {
   /// EUTRA Band
   uint16_t eutra_band;
   /// DL carrier frequency
-  uint32_t dl_CarrierFreq;
+  uint64_t dl_CarrierFreq;
   /// UL carrier frequency
-  uint32_t ul_CarrierFreq;
+  uint64_t ul_CarrierFreq;
   /// TX attenuation
   uint32_t att_tx;
   /// RX attenuation
